requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
pytest
httpx
structlog